pyproject.toml
[tool.poetry]
name = "co-pilot-mind"
version = "1.0.0"
description = "CoPilotMind é um sistema modular para aquisição, transcrição e armazenamento estruturado de comunicações simuladas entre um operador GCS (Ground Control Station) em um ambiente de suporte a voos, utilizando o Microsoft Flight Simulator 2020."
authors = ["ramonviana"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
anyio = "4.9.0"
build = "1.2.2.post1"
cachecontrol = "0.14.2"
certifi = "2025.1.31"
cffi = "1.17.1"
charset-normalizer = "3.4.1"
cleo = "2.1.0"
colorama = "0.4.6"
crashtest = "0.4.1"
datetime = "5.5"
distlib = "0.3.9"
dulwich = "0.22.8"
fastjsonschema = "2.21.1"
filelock = "3.18.0"
findpython = "0.6.3"
h11 = "0.14.0"
httpcore = "1.0.7"
httpx = "0.28.1"
idna = "3.10"
installer = "0.7.0"
jaraco-classes = "3.4.0"
jaraco-context = "6.0.1"
jaraco-functools = "4.1.0"
keyring = "25.6.0"
more-itertools = "10.6.0"
msgpack = "1.1.0"
packaging = "24.2"
pbs-installer = "2025.3.17"
pkginfo = "1.12.1.2"
platformdirs = "4.3.7"
poetry-core = "2.1.2"
pycparser = "2.22"
pyproject-hooks = "1.2.0"
pytz = "2025.2"
pywin32-ctypes = "0.2.3"
rapidfuzz = "3.12.2"
requests = "2.32.3"
requests-toolbelt = "1.0.0"
setuptools = "78.1.0"
shellingham = "1.5.4"
sniffio = "1.3.1"
sounddevice = "0.5.1"
soundfile = "0.13.1"
tomlkit = "0.13.2"
trove-classifiers = "2025.3.19.19"
typing-extensions = "4.13.0"
urllib3 = "2.3.0"
virtualenv = "20.30.0"
webrtcvad = "2.0.10"
zope-interface = "7.2"
zstandard = "0.23.0"
openai-whisper = { git = "https://github.com/openai/whisper.git" }
yt-dlp = "^2025.3.31"
torch = "2.5.1"
torchvision = "0.20.1"
torchaudio = "2.5.1"
python-slugify = "^8.0.4"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
